refactor!: Migrate `eth-json-rpc-middleware` to `JsonRpcEngineV2` (#7065)
Migrates `@metamask/eth-json-rpc-middleware` to the `JsonRpcEngineV2`
conventions in a total overhaul of the package. The semantics of each
middleware should be the same as before.

One potential behavioral difference is that request cloning within
middleware now occurs with `klona` instead of `klona/full`. This is
because `klona/full` copies property descriptors such that cloned
requests are immutable but not frozen in the `Object.isFrozen()` sense.
This is unlikely to make a difference in practice to us, since our
request objects have historically been "plain" objects.

As with all #6327-related work up to this point, `json-rpc-engine`
received minor refactors to facilitate the use of the new abstractions
_in situ_.

Incidentally, all lint rule exceptions have been removed and all lint
warnings and errors have been resolved for `eth-json-rpc-middleware`.

Several other packages are changed to support or in consequence of the
migration:
- `network-controller`
  - The `NetworkClient` is migrated to `JsonRpcEngineV2`.
- The `RpcService` now uses `Readonly<JsonRpcRequest>` internally for
request objects received via its `request()` method, since the "fetch"
middleware now passes it deeply frozen request objects directly.
- `message-manager` and `signature-controller`
- Minor changes to types in order to harmonize them with the new
conventions.