chore: Temporarily disable `@typescript-eslint/naming-convention`
The rule `@typescript-eslint/naming-convention` was already disabled
for most files, but it was set to `warn` in a few places still. It's
now completely disabled instead.

This was done to reduce the set of changes that are part of the
upcoming switch to error suppression (#6790), and also for consistency.
We still want to use this rule, but we can gradually begin enforcing it
in a more controlled manner using error suppression, across the entire
repository at once rather than selectively.