ci: Test scripts separately from packages
Currently we test top-level monorepo scripts every time we run tests
for a specific package. Instead they've now been moved to a separate
job that runs just once.

This should reduce the time it takes to run tests by a few seconds,
and it reduces noise when the script tests actually fail.