ci: Fix `check-release` action
The `check-release` action is consistently failing in the merge queue because
the `BASE_REF` environment variable is double-prefixed.

The `BASE_REF` environment variable has been updated to add a prefix for the
`pull_request.base.ref` input, but not for `merge_group.base_ref`.

Flags have also been added to ensure the step fails if we can't find a merge
base, just to simplify debugging if this fails again.