Keep network statuses up to date as requests are made
NetworkController has the ability to analyze an RPC endpoint and capture
its availability status (that is, whether the controller is able to make
successful requests to the endpoint). However, this step occurs
automatically only once, when the RPC endpoint's network is switched to,
and so any changes in status while the network is being used will not
be reflected in state. This problem can be mitigated by periodically
calling `lookupNetwork` manually, but this is awkward, and usage of this
method should be kept in check so as not to create too many requests.

Ideally, the controller should keep track of network status as requests
are made. This commit implements this change by hooking into events
published by the network client and added in a previous commit.