Move peer deps for controller/service packages to direct deps
Some controllers and services expect other controllers and services to
pre-exist so that they can communicate with them. Traditionally the way
that we have expressed this dependency is via peer dependencies. That
is, if a controller in `@metamask/foo-controller` used a service in
`@metamask/bar-service`, then it would declare a peer dependency on
a major version of `@metamask/bar-service`. This would enforce that
clients declare a direct dependency on `@metamask/bar-service` and the
major version used for this dependency matched the major version that
`@metamask/foo-controller` expressed.

However, this approach complicates the release process. If the major
version of a package is bumped, it must be synchronized across all peer
dependents across the monorepo. But as a result, all peer dependents
must be bumped by a major (since changing a peer dependency is a breaing
change) and thus must be included in the release. This means that
releases can quickly balloon in size, causing pain for other teams.

To address this issue, this commit converts all peer dependencies on
controller and service packages to regular direct dependencies. This
allows major versions of packages to be released more flexibly as needed.

However, this change also means that, at the client level, it is
possible for different versions of packages to be present in the
dependency tree. For instance, the client could directly depend on
`@metamask/foo-controller` 1.0.0 and `@metamask/bar-controller` 1.0.0,
but `@metamask/bar-controller` could itself depend on
`@metamask/foo-controller` 1.1.0. So while `@metamask/bar-controller`
thinks it is talking to a later version of `@metamask/foo-controller`,
it would actually be talking to an earlier version.

In most cases any incompatibilities would show up at compile-time, but
they could also show up at runtime. In general we expect client
engineers to be aware of these differences when making new releases —
perhaps more aware than they might be now.