chore: Use stricter TypeScript configuration for linting
The `tsconfig.json` file has been updated to use stricter compiler
options, matching what is used in other TSConfig files already. This
was a blocker for using the rule
`@typescript-eslint/no-unnecessary-boolean-literal-compare` with newer
versions of `typescript-eslint` (it requires `strictNullChecks` to work
correctly).

To avoid duplicating this configuration, it was moved to a new file
(`tsconfig.base.json`), and extended elsewhere. I've added a comment
block to each root-level `tsconfig` file explaining its purpose as
well, since it was getting difficult to keep track of them all.