chore: Update `typescript-eslint`
The ESLint plugin `typescript-eslint` (and related packages` have been
updated to use `^8.39` as the new minimum version. This is the minimum
now required on the `main` branch of `eslint-config` (not yet
released), as it includes a widening of `typescript` peer dependency.

This update unblocks the re-enabling of some disabled ESLint rules
(one of them had a bug in the auto-fixer that was fixed in v8.31.0).