chore: Re-emnable `@typescript-eslint/no-unnecessary-type-assertions`
The rule `@typescript-eslint/no-unnecessary-type-assertions` has been
re-enabled, and all lint violations fixed with `lint:fix`.

Some instances required more manual changes, I'll leave comments in the
PR for each case. Most of them were cases where a template literal was
casted to a type, which this rule regarded as unnecessary. In those
cases I added back an `as const` so that it didn't widen to the type
`string`.