chore: Improve `makeRequest` test utility function
The `makeRequest` utility function is now easier to control the return
type of using a type parameter, while still defaulting to the correct
type by default most of the time.

The previous version of this utility function frequently required a
type cast to prevent the return type from being resolved to `never`. It
was technically possible to specify the return type with a type
parameter, but it was onerous because the exact input type would also
be required as the first parameter.

This was done to simplify a later PR that enables a lint rule
preventing unnecessary type assertions. The rule ended up breaking
these tests because the cast impacted how the type was inferred in a
way that the rule didn't predict.