Fix getNetworkClientById to disregard format of network client ID
Currently, it is expected that IDs for Infura network clients match a
known property in the `InfuraNetworkType` enum and that IDs for custom
network clients follow a UUID format.

The way that these expectations are enforced is inconsistent: while
an Infura network client ID _must_ match an entry in the aforementioned
enum — an error will be produced upon initialization otherwise — there
is no such restriction for a custom network client ID.

But `getNetworkClientById` is more restrictive, and assumes that if
given a property of `InfuraNetworkType`, it _must_ correspond to an
Infura network client, failing if it cannot find one.

This causes issues with the `transaction-controller` package, because it
uses a flattened version of the network client registry to know
which networks should be used for tracking transactions. The flattened
network client registry just holds a list of network client IDs without
knowing which are Infura and which are custom. When
`transaction-controller` attempts to look them again,
`getNetworkClientById` throws an error. Since this operation occurs when
the client starts, this error ends up bricking the wallet.

By fixing `getNetworkClientById` so that network client IDs are treated
equally, we avoid the error. Incidentally, this change aligns with
plans to consolidate Infura and custom network clients in the future.