fix: Fix `get1559CompatibilityWithNetworkClientId`
This function assumed that metadata for a given network client would be
`undefined` if an EIP-1559 compatibility check was needed. This used to
be the case prior to `@metamask/network-controller@27.0.0`, but as of
that version we began regularly updating metadata without checking for
EIP-1559 compatibility.

The function `get1559CompatibilityWithNetworkClientI` now checks for
whether EIP-1559 compatibility metadata is present, not just whether
metadata is defined at all.d